requirements.txt
django == 5.1
djangorestframework==3.15.2
django-cors-headers==4.7.0